pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cartopy>=0.23.0",
    "db-dtypes>=1.3.0",
    "fastapi>=0.115.0",
    "fastapi-cache2>=0.2.2",
    "fastapi-pagination>=0.12.27",
    "fiona>=1.10.1",
    "google-cloud-bigquery>=3.25.0",
    "google-cloud-bigquery-storage>=2.26.0",
    "google-cloud-storage>=2.18.2",
    "infisical<1.6",
    "loguru>=0.7.2",
    "matplotlib>=3.9.2",
    "numpy>=2.1.1",
    "orjson>=3.10.7",
    "pandas>=2.2.2",
    "pendulum>=3.0.0",
    "redis>=5.0.8",
    "sentry-sdk[fastapi]>=2.14.0",
    "xarray>=2024.9.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff.format]
docstring-code-format = true
docstring-code-line-length = "dynamic"
indent-style = "space"
line-ending = "auto"
quote-style = "double"

[tool.taskipy.tasks]
serve = "uvicorn app.main:app --reload --port 8080"

[tool.uv]
dev-dependencies = [
    "taskipy>=1.13.0",
    "tomli>=2.0.1",
    "pre-commit>=3.8.0",
    "ruff>=0.6.5",
]
